Assert Symfony Console program's can be "expected"
This commit changes the buffer to mix stdout and stderr into one pseudo-terminal stream. This allows console programs to detect whether they are talking to a terminal or not, perhaps influencing interactivity (as is the case with the Symfony Console component).

Also, when setting an expectation, the buffer is first matched before selecting the stream for reading.